mm/migrate: pass migrate_detail around instead of plain 'reason'
From: Dave Hansen <dave.hansen@intel.com>

It would be ideal to keep some more information about the reasons
for migrations so that they can be analyzed.  There is more than
one reason to demote a page, for instance.

Introduce a new structure: 'struct migrate_detail' and pass it
around the migration paths instead of 'enum migrate_reason'.
Later, when more information is needed, it can be added to
the detail structure instead of changing all the call paths
again.