mm/numa: automatically generate node migration order
From: Dave Hansen <dave.hansen@linux.intel.com>

When memory fills up on a node, memory contents can be
automatically migrated to another node.  The biggest problems
are knowing when to migrate and to where the migration should
be targeted.

The most straightforward way to generate the "to where" list
would be to follow the page allocator fallback lists.  Those
lists already tell us if memory is full where to look next.
It would also be logical to move memory in that order.

But, the allocator fallback lists have a fatal flaw: most
nodes appear in all the lists.  This would potentially lead
to migration cycles (A->B, B->A, A->B, ...).

Instead of using the allocator fallback lists directly, keep
a separate node migration ordering.  But, reuse the same data
used to generate page allocator fallback in the first place:
find_next_best_node().

This means that the firmware data used to populate node
distances essentially dictates the ordering for now.

This code is unused for now.  It will be called later in the
series.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>

--

Changes since 2020/05/14 post:
 - Add additional commnts about READ/WRITE_ONCE() use
 - Add smb_wmb() to avoid exposing mixed before/after state
 - Fix comment typos/spelling