mm/numa: add node demotion data structure From: Dave Hansen <dave.hansen@linux.intel.com>
Prepare for the kernel to auto-migrate pages to other memory nodes
with a user defined node migration table. This allows creating single
migration target for each NUMA node to enable the kernel to do NUMA
page migrations instead of simply reclaiming colder pages. A node
with no target is a "terminal node", so reclaim acts normally there.
The migration target does not fundamentally _need_ to be a single node,
but this implementation starts there to limit complexity.

When memory fills up on a node, memory contents can be
automatically migrated to another node.  The biggest problems are
knowing when to migrate and to where the migration should be
targeted.

The most straightforward way to generate the "to where" list would
be to follow the page allocator fallback lists.  Those lists
already tell us if memory is full where to look next.  It would
also be logical to move memory in that order.

But, the allocator fallback lists have a fatal flaw: most nodes
appear in all the lists.  This would potentially lead to migration
cycles (A->B, B->A, A->B, ...).

Instead of using the allocator fallback lists directly, keep a
separate node migration ordering.  But, reuse the same data used
to generate page allocator fallback in the first place:
find_next_best_node().

This means that the firmware data used to populate node distances
essentially dictates the ordering for now.  It should also be
architecture-neutral since all NUMA architectures have a working
find_next_best_node().

The protocol for node_demotion[] access and writing is not
standard.  It has no specific locking and is intended to be read
locklessly.  This was done so that node_demotion[] locking has
no chance of becoming a bottleneck on large systems

Readers of node_demotion[] (such as next_demotion_node() callers)
must take care to avoid observing changes that appear incoherent.
For instance, even though no demotion cycles are allowed, it's
possible that a cycle could be observed.

Let's say that there are three nodes, A, B and C.  node_demotion[]
is set up to have this path:

	A -> B -> C

Suppose it was modified to instead represent this path:

	A -> C -> B

There is nothing to stop a reader from seeing B->C and then a
moment later seeting C->B.  That *appears* to be a cycle.  This
can be avoided with RCU and will be implemented in a later patch.

This code is unused for now.  It will be called later in the
series.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>
Reviewed-by: Yang Shi <shy828301@gmail.com>
Reviewed-by: Oscar Salvador <osalvador@suse.de>
Cc: Wei Xu <weixugc@google.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Huang Ying <ying.huang@intel.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: David Hildenbrand <david@redhat.com>

--

Changes from 20210331:
 * Merge patches for data structure definition and initialization

Changes from 20210302:
 * Fix typo in node_demotion[] comment

Changes since 20200122:
 * Make node_demotion[] __read_mostly
 * Add big node_demotion[] comment

changes in July 2020:
 - Remove loop from next_demotion_node() and get_online_mems().
   This means that the node returned by next_demotion_node()
   might now be offline, but the worst case is that the
   allocation fails.  That's fine since it is transient.