x86/microcode: Clear lead thread apply logic
From: Dave Hansen <dave.hansen@linux.intel.com>

SMT threads share a CPU core, including a microcode image.  Because
of this, only one thread within a core needs to apply microcode.

There is logic in __reload_late() to reflect this.  It ensures that
the lead thread applies the microcode first while the other threads
wait.  The opposing logic is implemented with two cpumask calculations:

       if (cpumask_first(topology_sibling_cpumask(cpu)) == cpu)
and
       if (cpumask_first(topology_sibling_cpumask(cpu)) != cpu)

This is perfectly functional.  The end effect is that the conditions
are mututally exclusive a CPU will fall into only one of the if()
blocks.

But, this construct is hard to understand.  The conditions look
very visually similar.  Instead of using mutually-exclusive cpumask
operations, introduce a variable which is set in the first if()
and then consulted at the second site.

This makes it perfectly clear that the later condition is paired
with this block:

	lead_thread = false;
	goto wait_for_siblings;

It also gives the cpumask calcuation a nice, symbolic name:
'lead_thread'.

No functional change.

Signed-off-by: Dave Hansen <dave.hansen@linux.intel.com>